backend/requirements.txt
flask
flask-cors
flask-sqlalchemy
flask-bcrypt
flask-migrate
flask-script
